# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Catalog API
  version: 1.0.0
  description: API для управління категоріями та товарами в каталозі.
servers:
  - url: http://localhost:8080/api
paths:
  /categories:
    get:
      summary: Отримати всі категорії
      description: Отримання списку категорій із можливістю фільтрації за назвою та пагінації.
      parameters:
        - name: page
          in: query
          description: Номер сторінки
          required: false
          schema:
            type: integer
        - name: size
          in: query
          description: Кількість елементів на сторінці
          required: false
          schema:
            type: integer
        - name: nameFilter
          in: query
          description: Фільтрація за назвою категорії
          required: false
          schema:
            type: string
      responses:
        200:
          description: Успішне отримання списку категорій
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Додати нову категорію
      description: Додає нову категорію до каталогу.
      requestBody:
        description: Дані нової категорії
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        201:
          description: Категорію створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /categories/{name}:
    get:
      summary: Отримати категорію
      description: Отримати дані однієї категорії за її назвою.
      parameters:
        - name: name
          in: path
          description: Назва категорії
          required: true
          schema:
            type: string
      responses:
        200:
          description: Успішне отримання категорії
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        404:
          description: Категорія не знайдена
    put:
      summary: Оновити категорію
      description: Оновлює інформацію про категорію за назвою.
      requestBody:
        description: Дані для оновлення категорії
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        200:
          description: Категорію оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        404:
          description: Категорія не знайдена
    delete:
      summary: Видалити категорію
      description: Видаляє категорію за назвою.
      responses:
        204:
          description: Категорія видалена
        404:
          description: Категорія не знайдена
components:
  schemas:
    Category:
      type: object
      properties:
        name:
          type: string
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
    Product:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number